Closes #91
* delete endpoint - deleteUser

* deleteOwner - Check if they have any hotels throw an error

* fix

+++ services/adminService.js
@@ -30,23 +30,14 @@ exports.acceptUserToOwner = async (id) => {
 }
 
 exports.deleteOwner = async (id) => {
+  const hotel = await Hotel.find({ ownerId: id })
+  if (hotel) throw new BadRequestError('Remove hotel(s) first')
   const user = await User.findOneAndDelete({
     _id: id,
     role: HOTEL_OWNER_ROLE,
   })
-
   if (!user) {
     throw new BadRequestError('Hotel owner with provided id not found')
-  }
-}
-
-exports.deleteUser = async (id) => {
-  const user = await User.findOneAndDelete({
-    _id: id,
-    role: USER_ROLE,
-  })
-  if (!user) {
-    throw new BadRequestError('User not found')
   }
 }
